Don't always reenable the transaction-update timer on update
When executing the update orchestration, it's not always desired to reenable the
transactional-update. Moreover, sometimes the timer was on a weird state if the
update failed. This commit refines the update orchestration following this
logic:

1. Check first which nodes have explicitely disabled the timer, so they are not
   reenabled in the end.
2. Disable the timer at the very beginning.
3. If this is a migration, reenable it always.
4. Otherwise, reenable the timer only for these nodes which did not explicitely
   disable it before the update.

Last but not least, I've also tweaked the `transactional-update/init.sls` file
so the timer is not touched by this file when updating.

bsc#1113518

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>