update.sls: start transactional-update timer
In the update process we disable the transactional update timer and re
enabled it after update the machine. However, just re enable it is not
enough. We need to start the timer. Otherwise, the update service will
not run and consequently no updates is installed. This commit adds in
the update.sls a step to start the timer after re enabling it.

bsc#1141675

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>